Move families from Element to Gedcom and improve
Moving "families" method to Gedcom class instead of Element class.
This is really more a property of the Gedcom object as a whole, not of
the Element object (as are any methods in Element that are relying on
the "dict" currently passed to it -- my goal is to remove this).
In other words, the operation should be:
gedcom_record.families(individual), not individual.families().

Also, for Element class, add method "is_family" and rename method
"individual" to "is_individual" (more descriptive). Use these to check
elements are correctly classified.

Finally, add functionality for returning FAMC families (where the
individual is a child) instead of FAMS families (where the individual
is a spouse).